feat(board): refactor KanbanBoard and introduce useKanbanScrollLogic for improved scrolling experience 🎉✨
